+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - `CanAdd(T item)`
 - `IStackInventory<T>` and `StackInventory<T>`
   - `CanAdd(T item)`
+  - `CanAddAt(T item, int index)`
 
 ## Known Issues
 * **The Current Architecture is not stable for the final version yet**